requirements.txt
torch>=2.0.0
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.21.0
openai>=1.0.0
pillow>=8.0.0
pyyaml>=6.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
scikit-learn>=1.0.0
requests>=2.25.0
python-dotenv>=0.19.0 